# Improve `ts_query_new` performance using pointer indirection