Profiling the `ts_query__analyze_patterns` function shows that it spends a lot of time copying items in its various state sets. These state sets are kept sorted, and the items are fairly large, so any time that we insert new entries near the front of the array, a lot of large calls to memcpy must occur.

While experimenting with more drastic ways to reduce this, I noticed an easy win along the way: by dynamically allocating each `AnalysisState`, we can reduce the size of the actual _array_ since the items become smaller (68 bytes vs 8 bytes). This gave me a 15-20% performance improvement on my [`experimental-tree-sitter-swift`](https://github.com/alex-pinkus/experimental-tree-sitter-swift) benchmarks. According to the benchmarks in this repository, performance is mostly flat, with the added pointer-chasing _not usually_ causing a problem (although there are a few very quick tests that got 1ms slower).

In addition to the pointer-chasing, one downside of the additional dynamic allocation is that we must manage the memory and be more careful to avoid sharing. To enforce proper usage, I modified `AnalysisStateSet` to _wrap_ an array, rather than just being a typedef for an array. The `AnalysisStateSet` always owns its contents, and creates a clone of any items that are requested of it, and the wrappers enforce that and call it out.

### Benchmarks on my own grammar
Here is the output of `cargo bench` on `experimental-tree-sitter-swift`, where I saw the 20% improvement:
_Before:_
```
Benchmarking highlights/ts_query_new: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 124.6s, or reduce sample count to 10.
highlights/ts_query_new time:   [1.2444 s 1.2451 s 1.2459 s]                                       
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
```
_After:_
```
Benchmarking highlights/ts_query_new: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 102.9s, or reduce sample count to 10.
highlights/ts_query_new time:   [1.0337 s 1.0350 s 1.0363 s]                                       
                        change: [-16.995% -16.878% -16.755%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  11 (11.00%) high mild
```

Comparing the PDFs of the two, we see that even the outlier data points with dynamic allocation are faster than without.

_Before_:
![current-master-branch](https://user-images.githubusercontent.com/62194897/148707675-bd3497fc-43f7-45ad-af0f-43a8e07a7270.png)
_After_:
![with-analysis-state-set-ptrs](https://user-images.githubusercontent.com/62194897/148707684-184a45a0-9297-4f51-915e-2d012151d9be.png)

To understand the reason this changed so much, compare the flamegraph visualizations for each:
_Before_:
![flamegraph-current-master-branch](https://user-images.githubusercontent.com/62194897/148708503-adf44a59-1491-4c49-9f0f-b8f1143ee9f2.png)

_After_:
![flamegraph-with-analysis-state-ptrs](https://user-images.githubusercontent.com/62194897/148708508-0ca8313e-15ef-4091-9c0a-bd490d07b5bf.png)

You can see that the current master branch spends a solid chunk of its time on `array_splice` and the corresponding `memmove`, and that once the data is hidden behind the smaller pointers, that time almost entirely goes away.

### Benchmarks from this repository
To get these numbers, I ran a modified version of the `tree-sitter` benchmarks that skips non-query-related tests and collects measurements for queries. I then upped the repetition count to 500 to get more granular data. Most of the performance numbers are identical; the differences are:
```
# < is the current master branch, > is the performance I see with these changes
$ diff current-master-branch.txt with-analysis-state-set-ptrs.txt
40c40
<     tags.scm                                      	time 5 ms	speed 324 bytes/ms
---
>     tags.scm                                      	time 6 ms	speed 278 bytes/ms
65,67c65,67
<     locals.scm                                    	time 32 ms	speed 27 bytes/ms
<     highlights.scm                                	time 54 ms	speed 43 bytes/ms
<     tags.scm                                      	time 106 ms	speed 10 bytes/ms
---
>     locals.scm                                    	time 40 ms	speed 21 bytes/ms
>     highlights.scm                                	time 49 ms	speed 47 bytes/ms
>     tags.scm                                      	time 101 ms	speed 10 bytes/ms
81c81
<   Average Speed (normal): 234 bytes/ms
---
>   Average Speed (normal): 232 bytes/ms
```

## Conclusion
**I'd like to merge some form of this change - it seems to be a solid improvement to me**. I'd appreciate guidance on additional validation. All unit tests still pass, of course, but what else should I be doing to make sure this change didn't break anything? @maxbrunsfeld  mentioned [on twitter](https://twitter.com/maxbrunsfeld/status/1479539497125617665) that there are other languages this is slow for - what other grammars should I benchmark to see if this yields improvements (and to validate that it doesn't cause regressions)?
And while this was some obvious low-hanging fruit, I think there are more drastic changes that can do even better. I'm curious for others' thoughts on these **future improvements**:
1. **Adding some fancier data structures so that the query analysis logic doesn't have to maintain these sorted lists.** If this were Rust, of course, we could use a `BTreeSet` and be done with it. Is there appetite for some additional data structures defined as macros and implemented locally? Have you previously tried that and found that the overhead isn't worth the benefit?
2. **Running this logic at grammar compile time.** It seems like a lot of repositories have their queries checked into the same repository, and that a lot of this sorting and finding paths is reasonable as a compilation step rather than a runtime step. Perhaps this info could be embedded in `parser.c` with some checksum, and on trying to load a query with a compatible checksum, we could speed up the load time by using the preloaded data structures?

Of course, I wouldn't want to block this PR on any of those future improvements - but thought I might bring them up for discussion.