@maxbrunsfeld 2022-01-10
Wow, awesome work @alex-pinkus, and thanks for the wonderful writeup! This looks a big improvement; I'll give a bit more careful review tomorrow, and will also be glad to discuss your suggested follow-up directions.

Off the top of my head, another language that exhibited slowdowns during query analysis was the [elixir grammar](https://github.com/elixir-lang/tree-sitter-elixir), as reported [here](https://github.com/nvim-treesitter/nvim-treesitter/issues/1957). I don't remember the details right now, but I think that when I was last reproducing the slowness for Elixir, I used the following query file and Elixir source file:

```clj
; query file
(call target: (identifier) @function)
```

```ex
# elixir file
quote do
  A.B.c()
end
```

Aside from that, I'd just want to make sure it doesn't slow down query construction for other languages, such as [the ones](https://github.com/tree-sitter/tree-sitter/blob/master/cli/src/tests/corpus_test.rs) that Tree-sitter uses as part of its test suite.

There's a benchmark in this repo that profiles query construction. Maybe we should add your Swift grammar and the official Elixir grammar to the set of grammars that Tree-sitter uses in its own CI and [benchmark](https://github.com/tree-sitter/tree-sitter/blob/master/cli/benches/benchmark.rs), which also exercises query construction.