@alex-pinkus 2022-01-10
It's a good thing you pointed me at `tree-sitter-elixir`—a quick microbenchmark had it performing _worse_ with this initial PR than without. Luckily, when I looked at the code with fresh eyes I realized that it would be very easy to bring in some basic object pooling for those `AnalysisState` items. With object pooling in place, I'm seeing a 15% reduction now for `tree-sitter-elixir` (1.475 secs with those two changes vs 1.738 secs without either) and a _28%_ for `experimental-tree-sitter-swift`!

New commit incoming that adds object pooling. Let me know if you need the commits squashed for hygiene.