@alex-pinkus 2022-01-10
Sure, we can share one object pool. My thinking was that even without the pool, the `AnalysisStateSet` would still be a struct. By having all usage go through the helper functions that enforce ownership invariants, we avoid getting into tricky states where two sets contain the same item and only one of them should free it. Accessing `states` directly and doing `array_*` operations on them becomes a code smell. That said, it’s your codebase — I’m happy to switch to the less-verbose approach if you feel strongly.

Will post an updated rev this evening!