@maxbrunsfeld 2022-01-10
I think it definitely helps clarify the ownership to have have all of the *writes* go through the helper functions; I just think that specifically for reading the *length* of the set, and for borrowing an entry from the array by index, I'm fine with reading the fields directly.