@alex-pinkus 2022-01-10
Yeah, my worry is that if it’s legal to _sometimes_ treat it as an array, it becomes harder to spot when someone _shouldn’t_ be treating it that way. Then again, I might be overthinking it :)