@alex-pinkus 2022-01-11
OK! Latest changes are up (I squashed them for posterity, since otherwise the history would add the struct stuff and then undo it). I ran `criterion` benchmarks for swift, elixir, kotlin (noticing @theHamsta's having tagged this issue), and cpp (for the other end of the spectrum, being a grammar that takes 20ms to compile its highlight queries):

_Swift_:
```
highlights/ts_query_new time:   [1.4395 s 1.4422 s 1.4452 s]
                        change: [-41.550% -41.065% -40.487%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 20 measurements (5.00%)
  1 (5.00%) high mild
```
![pdf-swift](https://user-images.githubusercontent.com/62194897/148870981-03cbc319-795f-4fc8-a301-ae16c19e2a4e.png)

_Kotlin_:
```
highlights/ts_query_new time:   [318.05 ms 329.40 ms 342.65 ms]
                        change: [-23.845% -20.696% -17.323%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 20 measurements (10.00%)
  2 (10.00%) high mild
```
![pdf-kotlin](https://user-images.githubusercontent.com/62194897/148870989-ef7f46a6-021e-4d89-a534-2a65abb60fd1.png)

_Elixir_:
```
highlights/ts_query_new time:   [801.62 ms 822.99 ms 848.83 ms]                                  
                        change: [-15.002% -12.698% -9.9846%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 20 measurements (5.00%)
  1 (5.00%) high mild
```
![pdf-elixir](https://user-images.githubusercontent.com/62194897/148871000-9c492b89-9f7e-468c-ab36-850fdfecefaf.png)


_C++_:
```
highlights/ts_query_new time:   [20.889 ms 21.138 ms 21.405 ms]                                    
                        change: [-8.1952% -6.2113% -4.2796%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 200 measurements (5.50%)
  11 (5.50%) high mild
```
![pdf-cpp](https://user-images.githubusercontent.com/62194897/148871013-17403c54-ec2e-4acc-a042-2c4e70a80899.png)


These results are not apples-to-apples with the ones in the original post because I'm on my Mac and ran the original ones from my Linux desktop. The `change:` lines for each are all relative to the runtime of the current `master` branch running on my Mac as well (of course).

Benchmark results from this repo are available at: [old.txt](https://github.com/tree-sitter/tree-sitter/files/7843120/old.txt)  [new.txt](https://github.com/tree-sitter/tree-sitter/files/7843123/new.txt) but mostly show no meaningful change from `master`.