@maxbrunsfeld 2022-01-11
In terms of next steps for optimizing this, I believe that there is room for optimization via some additional ahead-of-time computation, as you mentioned in your second suggestion.

There is definitely a lot of redundancy in the way that we walk the state graph. Many of the "starting" states for a given grammar rule have identical behavior, and we could save some work by *not* seeding the analysis with *all* of the starting states. Specifically, any parse state with the same [`core_id`](https://github.com/tree-sitter/tree-sitter/blob/bf210f0c9ec7931c1a5f639461495db240aac149/cli/src/generate/tables.rs#L49) is going to behave identically with respect to this analysis.

One thing I'd like to try is to add a new array to the generated parsers, which stores the `core_id` for each state. Then we could use this data during query analysis, when initializing the "subgraphs". We could refrain from including multiple starting states with the same `core_id`.

I actually explored this approach when I first added query analysis, but after some very cursory testing, I found that it didn't improve things enough to be worth including in the initial implementation. Every time, we add some piece of data to the generated parsers, we have to bump Tree-sitter's ABI version (that expresses compatibility between the runtime library and the generated parsers), which causes a headache for some users. But now, I'm thinking that maybe for *some* grammars (like Elixir, Swift, Kotlin), this data actually might make a bigger difference than it did in the grammars that I initially tested against. So I think it might be worth exploring this direction further.