@alex-pinkus 2022-01-12
Thank you for the pointers! That all makes a ton of sense. I hadn't thought about the ABI version bump but of course we would have to do that.

To make sure I'm understanding you correctly: we wouldn't bump `MIN_COMPATIBLE` version, just the `TREE_SITTER_LANGUAGE_VERSION`, and query analysis would look at all starting states if it didn't have `core_id` to use to dedup. But that would still be painful for some users who are on older `tree-sitter-cli` versions because once parsers get regenerated with the newer language version, they stop being able to use them? Or is there anything I'm overlooking?